Added a config param in initApp :fire:
This adds ability to configure app before mounting it to DOM.
The calling function will be given ability to manage global scope,
hence the developer can delegate app level logic like, auth initializtion
and global error handling here.
Updated the example app to demonstrate the same. Reduce lines of code
is testimony for cleaner implementation when such handling is done
which was made possible because of this config param addition. :v: